Refactor onramp module
Move code out of __init__

Improve this later, `utils` is always
a horrible module name.
